chore: use `Adw.MultiLayoutView` in video player controls
