--- data/ui/components/player/video/full.blp
@@ -1,87 +1,0 @@
-using Gtk 4.0;
-using Adw 1;
-
-template $FullVideoControls : Adw.Bin {
-  Adw.Clamp {
-    margin-end: 6;
-    margin-bottom: 6;
-    margin-start: 6;
-    valign: end;
-    maximum-size: 800;
-    tightening-threshold: 600;
-
-    Box {
-      orientation: horizontal;
-
-      styles [
-        "osd",
-        "toolbar",
-      ]
-
-      Box {
-        spacing: 6;
-        halign: center;
-
-        Button {
-          icon-name: "skip-backwards-10-symbolic";
-          action-name: "player.skip-backwards";
-
-          styles [
-            "flat",
-          ]
-        }
-
-        Button play_button {
-          icon-name: "media-playback-start-symbolic";
-          action-name: "player.play-pause";
-
-          styles [
-            "flat",
-          ]
-        }
-
-        Button {
-          icon-name: "skip-forward-10-symbolic";
-          action-name: "player.skip-forward";
-
-          styles [
-            "flat",
-          ]
-        }
-      }
-
-      Box {
-        orientation: horizontal;
-
-        Label progress_label {
-          label: "00:10";
-
-          styles [
-            "numeric",
-            "caption",
-          ]
-        }
-
-        $PlayerScale scale {}
-
-        Label duration_label {
-          label: "03:10";
-
-          styles [
-            "numeric",
-            "caption",
-          ]
-        }
-      }
-
-      MenuButton menu_button {
-        icon-name: "view-more-symbolic";
-        direction: up;
-
-        styles [
-          "flat",
-        ]
-      }
-    }
-  }
-}
