+++ data/ui/meson.build
@@ -18,8 +18,6 @@ blueprints = custom_target('blueprints',
     'components/player/now-playing/details.blp',
     'components/player/now-playing/view.blp',
     'components/player/video/controls.blp',
-    'components/player/video/full.blp',
-    'components/player/video/mini.blp',
     'components/player/video/view.blp',
     'components/player/video/volume-controls.blp',
     'components/player/full.blp',
